don't turn nulls into empty strings in bulk load
File destinations encode null fields the same way they encode empty strings.
For example, if the separator for the output is ',' then a null field will
look something like this in the output: ',,'

When this is subsequently bulk loaded into the (postgres) datawarehouse it is
interpreted as an empty string, not a null. This effectively bypasses any
non-null constraints in the database.

Since it's preferable to have errors thrown in the ETL, this change instructs
the bulk load to interpret ',,' as a null. This is only a default, so other
values can be set for the null string.

Another approach would be to have the file destination encode null fields
differently from empty strings. It seems better to have both empty strings and
nulls cast to nulls, so non-null constraints catch both as errors.

Depends on http://github.com/mainej/adapter_extensions/tree/1c9321052ebea4f17485b2d80a1f167738f28b15

Not sure if the tests pass with mysql.