cycle through dates (not times) when using DateDimensionBuilder
This is extremely important. Different systems handle daylight
savings differently.  So, on some systems

  >> Date.parse('2006-04-02').to_time.tomorrow.to_s(:db)
  => "2006-04-03 00:00:00"

as expected. But on other systems

  >> Date.parse('2006-04-02').to_time.tomorrow.to_s(:db)
  => "2006-04-02 23:00:00"

That means Time#tomorrow is a very bad way to cycle through dates.  A
better solution is Date#+(1).  This patch goes a bit further to just
cycle throw a Range of dates.