reverting commit 11d97d0cfd939d6f5577b3
The original commit was dealing with the case that the ordering
of column/row values wasn't what you would normally expect.  Now
that the entire selection and sorting of column/row values is
delegated to the dimension class, which allows specifications
for how each level of a hierarchy should be sorted, this is no
longer necessary.