queries on two dimensions with same column name returned nothing
Cube#query names them the same thing in the SQL, so they clobber each other
in the result hash.

This patch renames them columns in the sql, including the dimension table
name, so that the column names are always unique

It also allows queries where the row/column dimension/hierarchy are the same.
Not sure when you'd want this, but it's now possible.

This only fixes NoAggregate.  Not sure if it's broken in other
aggregation strategies, or if it is, how to fix it.