adding screens that execute after post-processing
The most common use case will be verifying that data in the datawarehouse
matches expectations from the production tables. This data will likely be
counts of rows, and sums of important values.

The existing screens can't do this for a variety of reasons. First, the
bulk-loads haven't happened, so you can't query the datawarehouse tables to
see what the post-ETL data will be. To remedy this, you could try to a) do
calculations on the pre-ETL tables, and b) combine that with data that's about
to be loaded by the ETL. That means you would need to keep running totals
somewhere within the ETL. That could quickly become complicated.

As one example, consider RowCountScreen. When running on a single control,
it's mildly useful. It reports how many rows are going to be added in this run
of the ETL. Even for a single control, that number can be hard to match. There
are a lot of variables determining it: :new_records_only, SCDs create extra
rows, after_read and before_write processors remove rows, etc. But when you
run RowCountScreen in cunjunction with a batch, it keeps a running total of
how many rows will be loaded by *every control in the batch*. It's almost
impossible to make any statements like "I expect that in this ETL 286 rows
will be added to the user dimension, account dimension and registration facts,
combined."

Note: these new screens are not useful for bailing out of a bulk load before
anything bad happens. They simply warn you when something bad has already
happened. Given how complicated it is to predict whether something bad will
happen, this is middle ground.