type 2 change wasn't preserving primary keys
Type 2 needs to expire old row. It does this by deleting directly from the db,
and scheduling a copied *expired* version of the old row for insert. But
add_virtuals was overwriting the old primary key with a new primary key. This
meant that the copied version had two changes: it was expired, AND it had a
different primary key. That would break all the existing foreign key
relationships to the old version of the row, which totally breaks SCD type 2
functionality.

Also noticed by Martin Honermeyer in
http://rubyforge.org/pipermail/activewarehouse-discuss/2008-January/000240.html