type 1 updates weren't copying primary key
Type 1 changes need to update an existing record.  To do this with
a bulk load, the existing record is deleted, and a new version is
inserted.  When this happens, the primary key of the original record
has to be propogated to the new record, so that foreign key
relationships are preserved.