removed CRCs from SCDs
The idea of CRCs is to avoid querying the existing data warehouse for old
versions of a SCD record. This is achieved by comparing the new (hashed) SCD
record with a cached (hashed) version. The idea is good, but in practice, the
old version is almost always loaded from the data warehouse. Additionally,
this CRC implementation introduced a bug where the datawarehouse failed to
pick up SCD changes in certain scenarios.

Until there is more need for SCD optimizations, I'm removing CRCs.

For reference...

The old version was reloaded in the following cases, when there was a CRC change:
  Type 1, to copy primary key, and delete old version
  Type 2, to expire old version (delete & reload w/ exp date)
The old version was reloaded in the following cases, when there was a CRC match:
  Type 1 and 2, to check if CRC recorded, but record never inserted
  Type 2, to treat changed non-scd fields as Type 1 change: copy primary key, delete old version

The bug happened if a) an SCD ETL was started, b) the change processing
succeeded (meaning all the CRCs were created/updated), c) the data load
failed, e) the ETL was re-run. In that case, the CRCs were all changed, so on
the second ETL run, it would look like no changes had been introduced, even
though the OLAP record doesn't match the datawarehouse record.